Base class for dali drivers had been created, and class for daliserver had been moved ot a separate class.
